+++ app/src/main/java/com/example/notecleanapp/domain/usecases/GetNoteByIdUseCase.kt
@@ -6,11 +6,11 @@ import javax.inject.Inject
 
 interface GetNoteByIdUseCase {
 
-    suspend fun getNoteById(id: Long): Note?
+    suspend fun getNoteById(id: Int): Note?
 
     class Base @Inject constructor(private val repository: NoteRepository) : GetNoteByIdUseCase {
 
-        override suspend fun getNoteById(id: Long): Note? {
+        override suspend fun getNoteById(id: Int): Note? {
             return repository.getNoteById(id)
         }
     }
